fix: gcs text decoding with iconv-lite (#259)
I ran into an issue where, when reading a simple csv, the chardet
returned the ISO-8859-1 encoding which toString could not handle. This
resulted in the MCP server being unable to read files from the bucket. I
believe it was looking for something like latin1. iconv-lite is a bit
more forgiving her when it comes to decoding.

Co-authored-by: Dorian Stubblefield <dstubblefield@users.noreply.github.com>